pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sourcemap-indexer"
version = "0.1.0"
description = "Codebase indexer: walks files into SQLite and enriches metadata via any LLM"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "Felipe Bambirra", email = "lipex360@gmail.com" }]
dependencies = [
  "typer>=0.12",
  "httpx>=0.27",
  "pyyaml>=6.0",
  "pathspec>=0.12",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-cov>=5",
  "ruff>=0.6",
  "mypy>=1.11",
  "vulture>=2.11",
  "bandit>=1.7",
  "pylint>=3.2",
  "build>=1.2",
  "types-PyYAML>=6.0",
]

[project.scripts]
sourcemap = "sourcemap_indexer.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/sourcemap_indexer"]

[tool.hatch.build.targets.wheel.shared-data]

[tool.hatch.build]
include = ["src/sourcemap_indexer/**"]

[tool.ruff]
line-length = 100
target-version = "py311"
cache-dir = ".ruff_cache"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RET", "C90", "S"]
ignore = ["S603", "S607", "S608"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]

[tool.ruff.lint.mccabe]
max-complexity = 5

[tool.mypy]
strict = true
python_version = "3.11"
cache_dir = ".mypy_cache"

[tool.pytest.ini_options]
addopts = "--cov=sourcemap_indexer --cov-report=term-missing --cov-fail-under=95"
testpaths = ["tests"]
pythonpath = ["src"]
cache_dir = ".pytest_cache"

[tool.coverage.run]
source = ["sourcemap_indexer"]
omit = ["*/tests/*", "*/.venv/*", "*/lib/*.py", "*/domain/repository.py"]

[tool.coverage.report]
exclude_lines = ["if __name__ == .__main__.:", "pragma: no cover"]
fail_under = 100
show_missing = true

[tool.pylint.messages_control]
disable = "all"
enable = ["C0103"]

[tool.pylint.basic]
argument-rgx   = "^([a-z_][a-z0-9_]{2,}|[ijk_]|id|pi)$"
variable-rgx   = "^([a-z_][a-z0-9_]{2,}|[ijk_]|id|pi)$"
attr-rgx       = "^([a-z_][a-z0-9_]{2,}|[ijk_]|id|pi)$"
function-rgx   = "^([a-z_][a-zA-Z0-9_]{2,}|[ijk_]|id|pi)$"
method-rgx     = "^([a-z_][a-zA-Z0-9_]{2,}|[ijk_]|id|pi)$"
module-rgx     = "^[a-z_][a-z0-9_]*$"
class-rgx      = "^_?[A-Z][a-zA-Z0-9]{2,}$"

[tool.bandit]
skips = ["B404", "B603", "B607", "B608"]
exclude_dirs = ["tests", ".venv"]
